Move FirebaseCrashlyticsLogExceptionTree into its own package
