+++ app/src/main/kotlin/com/hedvig/app/HedvigApplication.kt
@@ -13,7 +13,7 @@ import com.datadog.android.rum.RumMonitor
 import com.hedvig.android.hanalytics.android.tracking.ApplicationLifecycleTracker
 import com.hedvig.app.feature.settings.Theme
 import com.hedvig.app.feature.whatsnew.WhatsNewRepository
-import com.hedvig.app.util.FirebaseCrashlyticsLogExceptionTree
+import com.hedvig.app.util.firebase.FirebaseCrashlyticsLogExceptionTree
 import org.koin.android.ext.android.inject
 import timber.log.Timber
 
@@ -25,9 +25,7 @@ open class HedvigApplication : Application() {
   override fun onCreate() {
     super.onCreate()
     ProcessLifecycleOwner.get().lifecycle.addObserver(applicationLifecycleTracker)
-    Theme
-      .fromSettings(this)
-      ?.apply()
+    Theme.fromSettings(this)?.apply()
 
     whatsNewRepository.removeNewsForNewUser()
