+++ auth/auth-android/src/main/kotlin/com/hedvig/android/auth/android/AuthenticatedObserver.kt
@@ -31,7 +31,7 @@ class AuthenticatedObserver : DefaultLifecycleObserver {
     }
   }
 
-  override fun onPause(ownerr: LifecycleOwner) {
+  override fun onPause(owner: LifecycleOwner) {
     authObservingJob?.cancel()
     authObservingJob = null
   }
